+++ scrapy/extensions/closespider.py
