+++ scrapy/extensions/corestats.py
