+++ scrapy/extensions/debug.py
