+++ scrapy/extensions/httpcache.py
