+++ scrapy/extensions/memdebug.py
