+++ scrapy/extensions/spiderstate.py
