+++ scrapy/extensions/statsmailer.py
