+++ scrapy/templates/project/module/settings.py.tmpl
@@ -82,4 +82,4 @@ NEWSPIDER_MODULE = '$project_name.spiders'
 #HTTPCACHE_EXPIRATION_SECS=0
 #HTTPCACHE_DIR='httpcache'
 #HTTPCACHE_IGNORE_HTTP_CODES=[]
-#HTTPCACHE_STORAGE='scrapy.contrib.httpcache.FilesystemCacheStorage'
+#HTTPCACHE_STORAGE='scrapy.extensions.httpcache.FilesystemCacheStorage'
