+++ tests/py3-ignores.txt
@@ -100,7 +100,9 @@ scrapy/downloadermiddlewares/retry.py
 scrapy/downloadermiddlewares/httpproxy.py
 scrapy/downloadermiddlewares/cookies.py
 scrapy/contrib/statsmailer.py
+scrapy/extensions/statsmailer.py
 scrapy/contrib/memusage.py
+scrapy/extensions/memusage.py
 scrapy/commands/deploy.py
 scrapy/commands/bench.py
 scrapy/mail.py
